Move update related methods to UpdateMethods
