+++ gramjs/client/index.js
@@ -1,7 +1,8 @@
 const BaseClient = require('./BaseClient')
+const UpdateMethods = require('./UpdateMethods')
 const UserMethods = require('./UserMethods')
 const { mix } = require('../Helpers')
 
-class TelegramClient extends mix(BaseClient).with(UserMethods) {}
+class TelegramClient extends mix(BaseClient).with(UpdateMethods, UserMethods) {}
 
 module.exports = TelegramClient
